pyproject.toml
[tool.poetry]
name = "sugarjazy"
version = "0.6.2"
description = "Parse json logs output from uber-go/zap library nicely"
authors = ["Chmouel Boudjnah <chmouel@chmouel.com>"]
maintainers = ["Chmouel Boudjnah <chmouel@chmouel.com>"]
readme = "README.md"
license = "Apache-2.0"
keywords = ["logs", "tekton", "kubernetes"]
homepage = "https://github.com/chmouel/sugarjazy"
repository = "https://github.com/chmouel/sugarjazy"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]

include = [
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
python-dateutil = "^2.8.0"
pylint = "*"
mypy = "^0.942"
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
types-python-dateutil = "^2.8.10"

[tool.poetry.scripts]
sugarjazy = "sugarjazy.cli:main"

[tool.poetry.urls]
issues = "https://github.com/chmouel/sugarjazy/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
